site/pyproject.toml
[project]
name = "django-unicorn.com"
version = "0.1.0"
description = "Documentation of django-unicorn"
authors = [
    { name = "Adam Hill", email = "adamghill@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "django>=3.1.0",
    "python-dotenv>=0.14.0",
    "gunicorn>=20.0.4",
    "whitenoise>=5.1.0",
    "django-compressor>=2.4",
    "django-unicorn>=0.65.2",
    "django-cache-memoize>=0.1.7",
    "django-redis>=4.12.1",
    "httpx>=0.14.2",
    "sentry-sdk>=0",
    "django-sphinx-docs>=0.4.1",
    "django-fbv<1",
    "django-libsass>=0.9",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pytest",
    "pytest-django",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
